# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Runsight API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/workflows": {
      "get": {
        "tags": [
          "Workflows"
        ],
        "summary": "List Workflows",
        "operationId": "list_workflows_api_workflows_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Workflows"
        ],
        "summary": "Create Workflow",
        "operationId": "create_workflow_api_workflows_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workflows/{id}": {
      "get": {
        "tags": [
          "Workflows"
        ],
        "summary": "Get Workflow",
        "operationId": "get_workflow_api_workflows__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Workflows"
        ],
        "summary": "Update Workflow",
        "operationId": "update_workflow_api_workflows__id__put",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Workflows"
        ],
        "summary": "Delete Workflow",
        "operationId": "delete_workflow_api_workflows__id__delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          },
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Force"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowDeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workflows/{id}/commits": {
      "post": {
        "tags": [
          "Workflows"
        ],
        "summary": "Commit Workflow",
        "operationId": "commit_workflow_api_workflows__id__commits_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowCommitCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowCommitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workflows/{id}/simulations": {
      "post": {
        "tags": [
          "Workflows"
        ],
        "summary": "Create Workflow Simulation",
        "operationId": "create_workflow_simulation_api_workflows__id__simulations_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowSimulationCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowSimulationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workflows/{id}/enabled": {
      "patch": {
        "tags": [
          "Workflows"
        ],
        "summary": "Patch Workflow Enabled",
        "operationId": "patch_workflow_enabled_api_workflows__id__enabled_patch",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowEnabledUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workflows/{id}/regressions": {
      "get": {
        "tags": [
          "Workflows"
        ],
        "summary": "Get Workflow Regressions",
        "operationId": "get_workflow_regressions_api_workflows__id__regressions_get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CanvasViewport": {
        "properties": {
          "x": {
            "type": "number",
            "title": "X",
            "default": 0.0
          },
          "y": {
            "type": "number",
            "title": "Y",
            "default": 0.0
          },
          "zoom": {
            "type": "number",
            "title": "Zoom",
            "default": 1.0
          }
        },
        "type": "object",
        "title": "CanvasViewport"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WarningItem": {
        "properties": {
          "message": {
            "type": "string",
            "title": "Message"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "message"
        ],
        "title": "WarningItem"
      },
      "WorkflowCanvasState": {
        "properties": {
          "nodes": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Nodes"
          },
          "edges": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Edges"
          },
          "viewport": {
            "$ref": "#/components/schemas/CanvasViewport"
          },
          "selected_node_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Node Id"
          },
          "canvas_mode": {
            "type": "string",
            "enum": [
              "dag",
              "state-machine"
            ],
            "title": "Canvas Mode",
            "default": "dag"
          }
        },
        "type": "object",
        "title": "WorkflowCanvasState"
      },
      "WorkflowCommitCreate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "yaml": {
            "type": "string",
            "title": "Yaml"
          },
          "canvas_state": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkflowCanvasState"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "type": "string",
            "minLength": 1,
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "yaml",
          "message"
        ],
        "title": "WorkflowCommitCreate"
      },
      "WorkflowCommitResponse": {
        "properties": {
          "hash": {
            "type": "string",
            "title": "Hash"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "hash",
          "message"
        ],
        "title": "WorkflowCommitResponse"
      },
      "WorkflowCreate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "yaml": {
            "type": "string",
            "title": "Yaml"
          },
          "canvas_state": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkflowCanvasState"
              },
              {
                "type": "null"
              }
            ]
          },
          "commit": {
            "type": "boolean",
            "title": "Commit",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "yaml"
        ],
        "title": "WorkflowCreate"
      },
      "WorkflowDeleteResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "deleted": {
            "type": "boolean",
            "title": "Deleted"
          },
          "runs_deleted": {
            "type": "integer",
            "title": "Runs Deleted"
          }
        },
        "type": "object",
        "required": [
          "id",
          "deleted",
          "runs_deleted"
        ],
        "title": "WorkflowDeleteResponse"
      },
      "WorkflowEnabledUpdate": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          }
        },
        "type": "object",
        "required": [
          "enabled"
        ],
        "title": "WorkflowEnabledUpdate"
      },
      "WorkflowHealthMetrics": {
        "properties": {
          "run_count": {
            "type": "integer",
            "title": "Run Count",
            "default": 0
          },
          "eval_pass_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eval Pass Pct"
          },
          "eval_health": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "success",
                  "warning",
                  "danger"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Eval Health"
          },
          "total_cost_usd": {
            "type": "number",
            "title": "Total Cost Usd",
            "default": 0.0
          },
          "regression_count": {
            "type": "integer",
            "title": "Regression Count",
            "default": 0
          }
        },
        "type": "object",
        "title": "WorkflowHealthMetrics"
      },
      "WorkflowListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/WorkflowResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "WorkflowListResponse"
      },
      "WorkflowResponse": {
        "properties": {
          "kind": {
            "type": "string",
            "const": "workflow",
            "title": "Kind"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "yaml": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Yaml"
          },
          "canvas_state": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkflowCanvasState"
              },
              {
                "type": "null"
              }
            ]
          },
          "valid": {
            "type": "boolean",
            "title": "Valid",
            "default": true
          },
          "validation_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validation Error"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/WarningItem"
            },
            "type": "array",
            "title": "Warnings",
            "default": []
          },
          "block_count": {
            "type": "integer",
            "title": "Block Count",
            "default": 0
          },
          "modified_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modified At"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": false
          },
          "commit_sha": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commit Sha"
          },
          "health": {
            "$ref": "#/components/schemas/WorkflowHealthMetrics"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "id"
        ],
        "title": "WorkflowResponse"
      },
      "WorkflowSimulationCreate": {
        "properties": {
          "yaml": {
            "type": "string",
            "title": "Yaml"
          }
        },
        "type": "object",
        "required": [
          "yaml"
        ],
        "title": "WorkflowSimulationCreate"
      },
      "WorkflowSimulationResponse": {
        "properties": {
          "branch": {
            "type": "string",
            "title": "Branch"
          },
          "commit_sha": {
            "type": "string",
            "title": "Commit Sha"
          }
        },
        "type": "object",
        "required": [
          "branch",
          "commit_sha"
        ],
        "title": "WorkflowSimulationResponse"
      },
      "WorkflowUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "yaml": {
            "type": "string",
            "title": "Yaml"
          },
          "canvas_state": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkflowCanvasState"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "yaml"
        ],
        "title": "WorkflowUpdate"
      },
...
    }
  }
}